Separate SettlementRunCommand into command and job with functionality in SettlementRun::execute() Move Call Data Record get and parsing function to appropriate classes
